+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - The functions `permuteVariables` and `swapVariables` generated an error for a constant `qspray` or a `qspray` having only one term.
 
-- New function `isHomogeneous`, to check whether a `qspray` polynomial is homogeneous.
+- New function `isHomogeneousQspray`, to check whether a `qspray` polynomial is homogeneous.
 
 
 # qspray 3.0.0